Working(?) initial commit of new dotfile management system.
Similar to router system: for each detected template file, diffs against
the existing file and skips if no diff. If diff, prints the diff and
lets the user edit both the template file and the live file, to fix up
in either direction.

`click` turns out to be a really nice library for this basic UI stuff, I
was using `absl.flags` and `readchar` in an earlier draft.